cmds/exp/tcpdump: add icmp prints and filter
